Feedback: moving files, fixing comments and APIs
